Move more code into cascading-deletion.js
